requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
argcomplete==3.1.1
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blivet==3.8.0
blivet-gui==2.4.2
Brlapi==0.8.4
Brotli==1.0.9
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.0
click==8.1.3
cmake==3.26.1
comm==0.1.3
contourpy==1.0.7
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.11.1
cupshelpers==1.0
cycler==0.11.0
dasbus==1.7
dbus-python==1.3.2
debugpy==1.6.7
decorator==5.1.1
distro==1.7.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.4.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnf==4.18.0
dnspython==2.2.1
evdev==1.6.0
executing==1.2.0
facenet-pytorch==2.5.2
fedora-third-party==0.10
fer==22.5.0
file-magic==0.4.1
filelock==3.10.5
flatbuffers==23.3.3
fonttools==4.39.2
frozenlist==1.4.0
gast==0.4.0
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg==1.17.0
grpcio==1.53.0
gunicorn==22.0.0
h11==0.14.0
h5py==3.8.0
httpcore==1.0.5
httpx==0.27.0
humanize==3.13.1
idna==3.4
importlib-metadata==4.12.0
initial-setup==0.3.95
ipykernel==6.24.0
ipython==8.14.0
jaraco.classes==3.3.0
jax==0.4.7
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.12.0
keyring==23.11.0
kiwisolver==1.4.4
langtable==0.0.64
libclang==16.0.0
libcomps==0.1.20
libdnf==0.72.0
lightdm-gtk-greeter-settings==1.2.2
lit==16.0.0
lutris==0.5.14
lxml==4.9.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.0.4
moddb==0.8.1
more-itertools==8.12.0
mpmath==1.3.0
multidict==6.0.4
mutagen==1.46.0
mysql-connector-python==8.4.0
nest-asyncio==1.5.6
networkx==3.0
nftables==0.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
openai==1.31.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==9.2.0
platformdirs==3.9.1
ply==3.11
premailer==3.10.0
productmd==1.37
prompt-toolkit==3.0.39
protobuf==4.22.1
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.14.0
psutil==5.9.1
psycopg2-binary==2.9.9
ptyprocess==0.6.0
pure-eval==0.2.2
pwquality==1.4.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.20
pycryptodomex==3.18.0
pycups==2.0.1
pycurl==7.45.1
pydantic==2.7.3
pydantic_core==2.18.4
pyenchant==3.2.2
Pygments==2.15.1
PyGObject==3.42.2
PyJWT==2.8.0
pykickstart==3.41
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==3.0.9
pyparted==3.12.0
pypresence==4.3.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
pyrate-limiter==3.1.0
PySide2==5.13.2
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
PySocks==1.7.1
pyTelegramBotAPI==4.17.0
python-augeas==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-gettext==4.0
python-gnupg==0.5.0
python-manatools==0.0.4
python-meh==0.50
python-telegram-bot==21.1.1
pythondialog==3.5.3
pytz==2023.2
pyudev==0.23.2
pyxdg==0.27
PyYAML==6.0
pyzmq==25.1.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.1
rpm==4.18.2
rsa==4.9
scipy==1.10.1
SecretStorage==3.3.3
selinux==3.5
sepolicy==3.5
setools==4.4.0
setroubleshoot==3.3.32
shiboken2==5.13.2
shiboken6==6.6.1
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.0
sos==4.4
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
supervisor==4.2.5
sympy==1.11.1
systemd-python==235
telebot==0.0.5
telepot==12.7
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow @ file:///home/gilbert/Downloads/tensorflow-2.12.0-cp311-cp311-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tk==0.1.0
toolz==0.11.2
torch==2.0.0
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
twilio==8.7.0
typing_extensions==4.12.1
urllib3==1.26.18
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.14.1
yagmail==0.15.293
yarl==1.9.2
yt-dlp==2023.3.4
zipp==3.8.1
